Fix crash in replicationCacheMaster() expecting a nullptr cached_master
The replicationCacheMaster() function expects the master pointer to be
non-nullptr and the cached_master to be nullptr. But if we happened to
get disconnected multiple times and thus also go multiple times through
replicationCreateMasterClient(), then we end up with both master and
cached_master being non-nullptr, which then triggers an assertion in
replicationCacheMaster() which crashes the server.

When we are recreating the master struct, mark cached_master for
asynchronous freeing, and reset its pointer.

Fixes: #849
(cherry picked from commit 02a38501fd2a888dd2041ac49211cb399fa5acaa)